+++ web/src/recoil/stat.ts
@@ -9,6 +9,7 @@ export const activeTabState = atom<number>({
 
 export const statDate: IStatDateTypes = {
   activeDate: moment().format("YYYY.MM.DD"),
+  dateArray: [],
   startDate: "",
   endDate: "",
 };
