+++ web/src/types/stat.types.ts
@@ -1,5 +1,6 @@
 export interface IStatDateTypes {
   activeDate: string;
+  dateArray: string[];
   startDate: string;
   endDate: string;
 }
